% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDRUMLs.R
\name{BuildDRUMLs}
\alias{BuildDRUMLs}
\title{Build DRUML models for a list of drugs}
\usage{
BuildDRUMLs(
  df_input,
  .marker_database,
  df_response,
  drugs = NULL,
  input_type = "DRUML_model",
  save_path = "DRUML_models",
  save_csv = T,
  scale = T,
  computational_load = 0.8,
  partition.ratio = 0.7,
  models = c("glm", "cubist", "pls", "pcr", "nnet", "svm", "rf", "dl"),
  corr_method = "spearman",
  top_pos_n = 10,
  top_neg_n = 10,
  order_var = "r"
)
}
\arguments{
\item{df_input}{dataframe of input data.}

\item{.marker_database}{Markers database which will be used to for Drug Marker Enrichment}

\item{df_response}{Dataframe of drug responses(AAC) with cell lines as row names and column names as drugs.}

\item{drugs}{list of drugs from which models will be generated. If left NULL models will be built for all drugs in the marker database.}

\item{save_path}{save location for models and results csvs}

\item{save_csv}{if TRUE csv files containing DRUML results will be made}

\item{computational_load}{Set as a decimal fraction of the number of cores you which to be recruited for model building e.g a value of 0.8 means 80\% of all available CPU cores will be used to build models. If left as NULL only 1 core will be used.}

\item{models}{list of models to use for machine learning:
\itemize{
  \item glm = Baysian GLM model
  \item cubist = Cubist model
  \item pls = Partial list squares model
  \item pcr = Principal components regression
  \item nnet = Neural network model
  \item scm = Super vector machines
  \item rf = Random forests regressor model
  \item dl = Deep learning model using h2o package (needs Java to be up to date)
  }}

\item{top_pos_n}{maximum ammount of sensitive distance markers for DRUML models}

\item{top_neg_n}{maximum ammount of resistant distance markers for DRUML models}

\item{order_var}{variable used to filter Distance models for DRUML:
\itemize{
 \item p = p value
 \item r = r value
 \item slope = slope of fitted linear model
 }}

\item{method}{Statistical method for filtering distance markers for DRUML models default is spearman}
}
\description{
Title
}
