% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildEMDR.R
\name{BuildEMDR}
\alias{BuildEMDR}
\title{Empirical Markers of Drug Response}
\usage{
BuildEMDR(
  df_input = NULL,
  df_response = NULL,
  drugs,
  p.cutoff = 0.05,
  fold.cut.off = 1,
  resampling.times = 5,
  return_limma = F,
  computational_load = NULL,
  nfolds = 10,
  maxmarker_res = NULL,
  maxmarker_sens = NULL
)
}
\arguments{
\item{df_input}{input data with cell lines as colnames and varaibles as rownames}

\item{df_response}{response values dataframe with colnames as cell lines and drugs as column}

\item{p.cutoff}{P value threshold to define a repeat comparison as significant}

\item{fold.cut.off}{fold value threshold to define a repeat comparison direction}

\item{resampling.times}{number of resampling times}

\item{return_limma}{Set as TRUE if for limma analysis results to be returned}

\item{computational_load}{Set as a decimal fraction of cores you want to use. If left as NULL only 1 core will be used.}

\item{drug}{drug generate markers for}

\item{scale_input}{Set as TRUE to scale input data}
}
\description{
Empirical Markers of Drug Response
}
