% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModels.R
\name{.BuildModel}
\alias{.BuildModel}
\title{Internal Build model function}
\usage{
.BuildModel(
  df_train,
  df_test,
  model_type,
  drug,
  input_type,
  save_path,
  metric = "RMSE",
  initiate_h2o = T
)
}
\arguments{
\item{model_type}{string of model type to build e.g cubist, glm, nnet, pcr, pls, rf and svm}

\item{drug}{drug to use for naming model}

\item{input_type}{input type to use for naming model}

\item{save_path}{save file location}

\item{metric}{error metric to use to build model}

\item{df_input}{dataframe to use for model training. Response value must be in column named response.}

\item{partition.ratio}{partition ratio for making testing dataset}
}
\description{
Title Internal function for building machine learning models
}
\examples{
.BuildModel(df_input = df_distances, model_type = "rf", drug = "ABT-199", save_path = "my_model_dir")
}
