% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugMarkerEnrichment.R
\name{DrugMarkerEnrichment}
\alias{DrugMarkerEnrichment}
\title{Generate D values from input data and marker database}
\usage{
DrugMarkerEnrichment(df = df.ppindex, marker_database = "aml_prot_markers", scale = T, output = c("Distance", "zscore"))
}
\arguments{
\item{df}{input data with cell lines as colnames and varaibles as rownames}

\item{marker_database}{markers database to be used can be the following:
\itemize{
  \item phospho aml
  \item phospho solid_markers
  \item prot aml
  \item prot solid
  \item rna aml
  \item rna solid
  }
 @param marker_database_path path of the directory containing the DRUMLR marker file}

\item{scale}{Set as TRUE to scale input data}

\item{output}{The output metric you want to use can be "Distance" or "zscore"}
}
\description{
Get Distance values or zscores between sensitive and resistant markers
}
