% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSensitivity.R
\name{FilterSensitivity}
\alias{FilterSensitivity}
\title{Filter Drugs by sensitivity data}
\usage{
FilterSensitivity(
  sensitivity_database = NULL,
  iqr_tolerance = 0.04,
  n_coverage = 0.8,
  cell_names = NULL
)
}
\arguments{
\item{sensitivity_database}{sensitvitiy data with cell lines as rownamed and drugs as columns}

\item{iqr_tolerance}{IQR range cut off to filter drug list}

\item{n_coverage}{n_coverage cut off required to ensure models are reliable}

\item{cell_names}{List of cell names to subset sensitivity data}
}
\description{
Filter Drugs by sensitivity data
}
