% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinMaxNormalise.R
\name{MinMaxNormalise}
\alias{MinMaxNormalise}
\title{Normalise data by min max scaler}
\usage{
MinMaxNormalise(df = DRUMLRdata:::DRUML_phos, .margin=2)
}
\arguments{
\item{x}{dataframe or matrix to be scaled (all data must be numeric)}

\item{margin}{can be 1 or 2. x will be scaled by row if .margin =1 or  column .margin=2}
}
\description{
Normalise data by min max scaler
}
