% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimiseMLInput.R
\name{OptimiseMLInput}
\alias{OptimiseMLInput}
\title{Optimise ML input distance markers}
\usage{
OptimiseMLInput(
  df_train,
  method = "spearman",
  max_n_D = 30,
  min_n_D = 7,
  partition.ratio = 0.7,
  return_corr_analysis = FALSE,
  order_var = "r"
)
}
\arguments{
\item{df_train}{training dataset for machine learning. df_input with response variables in a column named response}

\item{method}{statistical method used to determine correlation}

\item{max_n_D}{maximum number of markers to use}

\item{min_n_D}{Minimum number of  markers to use}

\item{return_corr_analysis}{TRUE results of correlation analysis will be returned in addition to markers lists}

\item{order_var}{variable used to order which markers to use}
}
\description{
Title
}
