% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRUMLPredict_example.R
\name{PredictDRUML_example}
\alias{PredictDRUML_example}
\title{Make a heatmap of marker enrichment}
\usage{
PredictDRUML(markers = GetMarkers("barasetib")$sensitive$.,inputdata = LoadData("PDBphos"),tissuefilter = "haematopoietic_and_lymphoid_tissue",metric = "aac", drug = "ABT-199")
}
\arguments{
\item{input_type}{Input data type/model label}

\item{drugs}{Drug for which predictions will be made}

\item{models}{list of models for which you would like to use for predictions}

\item{shut_h2o}{if TRUE the h2o will be shutdown after predictions are made}

\item{df_input}{input dataframe}

\item{.scale}{if TRUE data will be scaled}

\item{model_path_file}{path to model path file which is created in model dir by build models}

\item{.marker_database}{markers database which will be used}
}
\description{
Make a heatmap of marker enrichment
}
