% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveRepeatNo.R
\name{RemoveRepeatNo}
\alias{RemoveRepeatNo}
\title{Update Internal Databases for KSEA}
\usage{
RemoveRepeatNo(colnames(df_input))
}
\arguments{
\item{repeatseperator}{The string which separates cell name from repeat number. For DRUML data "__" is used as default.}
}
\description{
Removes the repeat numbers for triplicate data so that datasets can be filtered by cell name
}
